Make more intelligent decisions about which calculations to do.
Previously, we tried every recursion relation all the time, using
exceptions to indicate what does and doesn't work. Now, we use various
conditions to tell ahead of time which calculations will succeed and
which will not.

This mostly just makes cases which rely on the exchange relations a
little faster.